+++ tester.js
@@ -1,9 +1,9 @@
 
 import { MangaSearch, Manga } from './mangaScraper/mangaScraper.js';
-import { Mangatoto } from './mangaScraper/Fetchers/mangatoto.js';
 
+let mangaSearch = new MangaSearch();
 let manga = new Manga();
-let mangatoto = new Mangatoto();
+
 
 //set config
 manga.updateConfig({
@@ -31,4 +31,15 @@ p.innerHTML = titles;
 let img = document.getElementById("2");
 img.src = pictureLinks[1][0];
 
-console.log(await mangatoto.search("hi"));
+console.log(await mangaSearch.search("hi"));
+
+function test(input) {
+    console.log(input);
+    return;
+}
+
+await mangaSearch.search("hi", test);
+
+mangaSearch.updateConfig({runCallbackOnError : true});
+
+mangaSearch.search("hi", test);